requirements.txt
Package                              Version
------------------------------------ -------
action-msgs                          2.0.2
action-tutorials-interfaces          0.33.5
action-tutorials-py                  0.33.5
actionlib-msgs                       5.3.5
aiohappyeyeballs                     2.4.3
aiohttp                              3.10.10
aiohttp-jinja2                       1.6
aioserial                            1.3.1
aiosignal                            1.3.1
ament-cmake-test                     2.5.2
ament-copyright                      0.17.1
ament-cppcheck                       0.17.1
ament-cpplint                        0.17.1
ament-flake8                         0.17.1
ament-index-python                   1.8.1
ament-lint                           0.17.1
ament-lint-cmake                     0.17.1
ament-package                        0.16.3
ament-pep257                         0.17.1
ament-uncrustify                     0.17.1
ament-xmllint                        0.17.1
angles                               1.16.0
attrs                                24.2.0
builtin-interfaces                   2.0.2
click                                8.1.7
composition-interfaces               2.0.2
cv-bridge                            4.1.0
demo-nodes-py                        0.33.5
diagnostic-msgs                      5.3.5
domain-coordinator                   0.12.0
example-interfaces                   0.12.0
examples-rclpy-executors             0.19.4
examples-rclpy-minimal-action-client 0.19.4
examples-rclpy-minimal-action-server 0.19.4
examples-rclpy-minimal-client        0.19.4
examples-rclpy-minimal-publisher     0.19.4
examples-rclpy-minimal-service       0.19.4
examples-rclpy-minimal-subscriber    0.19.4
frozenlist                           1.5.0
geometry-msgs                        5.3.5
idna                                 3.10
image-geometry                       4.1.0
interactive-markers                  2.5.4
Jinja2                               3.1.4
laser-geometry                       2.7.0
launch                               3.4.2
launch-ros                           0.26.5
launch-testing                       3.4.2
launch-testing-ros                   0.26.5
launch-xml                           3.4.2
launch-yaml                          3.4.2
lifecycle-msgs                       2.0.2
linkify-it-py                        2.0.3
logging-demo                         0.33.5
map-msgs                             2.4.1
markdown-it-py                       3.0.0
MarkupSafe                           3.0.2
mdit-py-plugins                      0.4.2
mdurl                                0.1.2
message-filters                      4.11.2
mpremote                             1.24.0
msgpack                              1.1.0
multidict                            6.1.0
nav-msgs                             5.3.5
osrf-pycommon                        2.1.4
pcl-msgs                             1.0.0
pendulum-msgs                        0.33.5
pip                                  24.0
platformdirs                         4.3.6
plotext                              5.3.2
propcache                            0.2.0
Pygments                             2.18.0
pyserial                             3.5
python-qt-binding                    2.2.1
qt-dotgraph                          2.7.4
qt-gui                               2.7.4
qt-gui-cpp                           2.7.4
qt-gui-py-common                     2.7.4
quality-of-service-demo-py           0.33.5
rcl-interfaces                       2.0.2
rclpy                                7.1.2
rcutils                              6.7.2
resource-retriever                   3.4.3
rich                                 13.9.4
rmw-dds-common                       3.1.0
ros2action                           0.32.1
ros2bag                              0.26.5
ros2bag-mcap-cli                     0.26.5
ros2bag-sqlite3-cli                  0.26.5
ros2cli                              0.32.1
ros2component                        0.32.1
ros2doctor                           0.32.1
ros2interface                        0.32.1
ros2launch                           0.26.5
ros2lifecycle                        0.32.1
ros2multicast                        0.32.1
ros2node                             0.32.1
ros2param                            0.32.1
ros2pkg                              0.32.1
ros2run                              0.32.1
ros2service                          0.32.1
ros2topic                            0.32.1
rosbag2-interfaces                   0.26.5
rosbag2-py                           0.26.5
rosgraph-msgs                        2.0.2
rosidl-adapter                       4.6.4
rosidl-cli                           4.6.4
rosidl-cmake                         4.6.4
rosidl-generator-c                   4.6.4
rosidl-generator-cpp                 4.6.4
rosidl-generator-py                  0.22.0
rosidl-generator-type-description    4.6.4
rosidl-parser                        4.6.4
rosidl-pycommon                      4.6.4
rosidl-runtime-py                    0.13.1
rosidl-typesupport-c                 3.2.2
rosidl-typesupport-cpp               3.2.2
rosidl-typesupport-fastrtps-c        3.6.0
rosidl-typesupport-fastrtps-cpp      3.6.0
rosidl-typesupport-introspection-c   4.6.4
rosidl-typesupport-introspection-cpp 4.6.4
rpyutils                             0.4.1
rqt-action                           2.2.0
rqt-bag                              1.5.4
rqt-bag-plugins                      1.5.4
rqt-console                          2.2.1
rqt-graph                            1.5.4
rqt-gui                              1.6.0
rqt-gui-py                           1.6.0
rqt-msg                              1.5.1
rqt-plot                             1.4.0
rqt-publisher                        1.7.2
rqt-py-common                        1.6.0
rqt-py-console                       1.2.2
rqt-reconfigure                      1.6.2
rqt-service-caller                   1.2.1
rqt-shell                            1.2.2
rqt-srv                              1.2.2
rqt-topic                            1.7.2
sensor-msgs                          5.3.5
sensor-msgs-py                       5.3.5
service-msgs                         2.0.2
shape-msgs                           5.3.5
sros2                                0.13.2
statistics-msgs                      2.0.2
std-msgs                             5.3.5
std-srvs                             5.3.5
stereo-msgs                          5.3.5
teleop-twist-keyboard                2.4.0
textual                              0.85.2
textual-dev                          1.6.1
textual-plotext                      0.2.1
textual-serve                        1.1.1
tf2-geometry-msgs                    0.36.4
tf2-kdl                              0.36.4
tf2-msgs                             0.36.4
tf2-py                               0.36.4
tf2-ros-py                           0.36.4
tf2-sensor-msgs                      0.36.4
tf2-tools                            0.36.4
topic-monitor                        0.33.5
trajectory-msgs                      5.3.5
turtle-tf2-py                        0.5.0
turtlesim                            1.8.3
type-description-interfaces          2.0.2
typing_extensions                    4.12.2
uc-micro-py                          1.0.3
unique-identifier-msgs               2.5.0
visualization-msgs                   5.3.5
yarl                                 1.17.1
